app/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
torch==2.2.0
torchvision==0.17.0
pillow==9.5.0
python-multipart==0.0.9
numpy==1.24.4
